refactor: structure authorization files into folders
